Day 13
Why did the assignment even say to find the cheapest way to solve each
machine? There's always exactly one way to solve each machine.

The only way there could be more than one solution is for the A and B
vectors to be linearly dependent - and such a case doesn't appear in the
input.

It would be fairly easy to Haskell this with a linear algebra library,
but I don't feel like setting up a whole cabal project to get one
library. So use Python with numpy.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>